message list: Collapse the MessageList / MessageListWeb distinction.
This indirection was important when we had two different
implementations of the message list, one with WebView and one within
React Native.  Now that we're firmly on the WebView version, the
distinction doesn't serve a clear purpose; we haven't maintained much
of a principled line about what goes on which side, and it's not clear
what a useful such line would be.

So, having made MessageList more and more vestigial in the preceding
commits until it was nothing but the `connect` call and friends, merge
that into MessageListWeb.

This commit moves a bunch of code verbatim, with minimal other changes
for clarity's sake.
